59. print any thing.
Statement:Polymorphism word always refer, a single thing can perform many actions or operations. So due to inspiration of polymorphism

sumit wants to print any thing using a single method display(). He wants to design display() such that it can print any thing what ever a programmer

wants to print with it. He has a great idea but don’t know how to implement it. In order to solve this problem he needs your help to design display() method

in polymorphic way.

For this create a class named as Print and define method display() such that it accepts anything and can display the received values.

for better understanding we are providing you a code snippet, this code snippet contains a main() method and some statements.

We have already written this code snippet in this program so there is no need to write main() method for this program.

So your task is to just observer the code snippet and write the rest code (class creation) which is required for successful execution of the program.


int main()
{
Print p;
p.display(100);
p.display(32768);
p.display(‘A’);
p.display(‘1’);
p.display(“C is Fun”);
p.display(3.333333333335);
p.display(8.5);
return 0;
}

So code the display() method such that it can print exact same thing that are passed to it ,each in new line.

Input: there is no need to provide any input, that are already provided by us.

Output: output should be exactly same as given below.

100

32768

A

1

C is Fun

3.333333333335

8.5

Constraint :don’t write main() method in your program, just complete the coding of class.